[ADD]各メタデータのパーサを追加
