+++ app/src/main/java/com/ddbj/ld/parser/BioSampleParser.java
@@ -1,81 +1,61 @@
 package com.ddbj.ld.parser;
 
+import javax.xml.stream.XMLInputFactory;
+import javax.xml.stream.XMLStreamConstants;
+import javax.xml.stream.XMLStreamException;
+import javax.xml.stream.XMLStreamReader;
 import java.io.BufferedInputStream;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
-import javax.xml.stream.XMLInputFactory;
-import javax.xml.stream.XMLStreamConstants;
-import javax.xml.stream.XMLStreamReader;
-import javax.xml.stream.XMLStreamException;
 
-import com.ddbj.ld.bean.BioProjectBean;
+import com.ddbj.ld.bean.BioSampleBean;
 
-public class BioProjectParser {
-    public static List<BioProjectBean> parse(String xmlFile) throws FileNotFoundException, XMLStreamException {
+public class BioSampleParser {
+    public static List<BioSampleBean> parse(String xmlFile) throws FileNotFoundException, XMLStreamException {
         XMLInputFactory factory = XMLInputFactory.newInstance();
         BufferedInputStream stream = new BufferedInputStream(new FileInputStream(xmlFile));
         XMLStreamReader reader = factory.createXMLStreamReader(stream);
-        int packageFlg = 0;
-        int descFlg = 0;
 
-        List<BioProjectBean> bioProjectBeanList = new ArrayList<>();
-        BioProjectBean bioProjectBean = null;
+        boolean isStarted = false;
+        boolean isDescription = false;
+        BioSampleBean bioSampleBean = null;
+        List<BioSampleBean> bioSampleBeanList = new ArrayList<>();
 
+        // TODO description
         for (; reader.hasNext(); reader.next()) {
             int eventType = reader.getEventType();
 
-            if (packageFlg == 0
+            if (isStarted == false
             && eventType == XMLStreamConstants.START_ELEMENT
-            && reader.getName().toString().equals("Package")) {
-                packageFlg = 1;
-                bioProjectBean = new BioProjectBean();
-            } else if (packageFlg == 1
-                    && eventType == XMLStreamConstants.START_ELEMENT
-                    && reader.getName().toString().equals("ArchiveID")) {
-                bioProjectBean.setIdentifier(parseArchiveID(reader));
-            } else if (packageFlg == 1
-                    && eventType == XMLStreamConstants.START_ELEMENT
-                    && reader.getName().toString().equals("ProjectDescr")) {
-                descFlg = 1;
-            } else if (descFlg == 1
-                    && eventType == XMLStreamConstants.START_ELEMENT
-                    && reader.getName().toString().equals("Name")) {
-                bioProjectBean.setName(reader.getElementText());
-            } else if (descFlg == 1
-                    && eventType == XMLStreamConstants.START_ELEMENT
-                    && reader.getName().toString().equals("Title")) {
-                bioProjectBean.setTitle(reader.getElementText());
-            } else if (descFlg == 1
+            && reader.getName().toString().equals("BioSample")) {
+                isStarted = true;
+                bioSampleBean = new BioSampleBean();
+                bioSampleBean.setIdentifier(AccessionParser.parseAccession(reader));
+            } else if (isStarted == true
                     && eventType == XMLStreamConstants.START_ELEMENT
                     && reader.getName().toString().equals("Description")) {
-                bioProjectBean.setDescription(reader.getElementText());
-                descFlg = 0;
-            } else if (packageFlg == 1
+                isDescription = true;
+            } else if (isDescription == true
+                    && eventType == XMLStreamConstants.START_ELEMENT
+                    && reader.getName().toString().equals("SampleName")) {
+                bioSampleBean.setName(reader.getElementText());
+            } else if (isDescription == true
+                    && eventType == XMLStreamConstants.START_ELEMENT
+                    && reader.getName().toString().equals("Title")) {
+                bioSampleBean.setTitle(reader.getElementText());
+            } else if (isStarted == true
                     && eventType == XMLStreamConstants.END_ELEMENT
-                    && reader.getName().toString().equals("Package")) {
-                packageFlg = 0;
-                bioProjectBeanList.add(bioProjectBean);
+                    && reader.getName().toString().equals("BioSample")) {
+                isStarted = false;
+                isDescription = false;
+                bioSampleBeanList.add(bioSampleBean);
             }
         }
 
         reader.close();
 
-        return bioProjectBeanList;
-    }
-
-    private static String parseArchiveID(XMLStreamReader reader) {
-
-        String accession = null;
-        int count = reader.getAttributeCount();
-        for (int i=0; i<count; i++) {
-            if (reader.getAttributeName(i).toString().equals("accession")) {
-                accession = reader.getAttributeValue(i);
-                break;
-            }
-        }
-
-        return accession;
+        return bioSampleBeanList;
     }
 }
