+++ app/src/main/java/com/ddbj/ld/parser/SampleParser.java
@@ -1,7 +1,5 @@
 package com.ddbj.ld.parser;
 
-import com.ddbj.ld.bean.BioSampleBean;
-
 import javax.xml.stream.XMLInputFactory;
 import javax.xml.stream.XMLStreamConstants;
 import javax.xml.stream.XMLStreamException;
@@ -12,49 +10,46 @@ import java.io.FileNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class BioSampleParser {
-    public static List<BioSampleBean> parse(String xmlFile) throws FileNotFoundException, XMLStreamException {
+import com.ddbj.ld.bean.SampleBean;
+
+public class SampleParser {
+    public static List<SampleBean> parse(String xmlFile) throws FileNotFoundException, XMLStreamException {
         XMLInputFactory factory = XMLInputFactory.newInstance();
         BufferedInputStream stream = new BufferedInputStream(new FileInputStream(xmlFile));
         XMLStreamReader reader = factory.createXMLStreamReader(stream);
 
         boolean isStarted = false;
-        boolean isDescription = false;
-        BioSampleBean bioSampleBean = null;
-        List<BioSampleBean> bioSampleBeanList = new ArrayList<>();
+        SampleBean sampleBean = null;
+        List<SampleBean> sampleBeanList = new ArrayList<>();
 
-        // TODO description
+        // TODO name
         for (; reader.hasNext(); reader.next()) {
             int eventType = reader.getEventType();
 
             if (isStarted == false
             && eventType == XMLStreamConstants.START_ELEMENT
-            && reader.getName().toString().equals("BioSample")) {
+            && reader.getName().toString().equals("SAMPLE")) {
                 isStarted = true;
-                bioSampleBean = new BioSampleBean();
-                bioSampleBean.setIdentifier(AccessionParser.parseAccession(reader));
+                sampleBean = new SampleBean();
+                sampleBean.setIdentifier(AccessionParser.parseAccession(reader));
+            } else if (isStarted == true
+                    && eventType == XMLStreamConstants.START_ELEMENT
+                    && reader.getName().toString().equals("TITLE")) {
+                sampleBean.setTitle(reader.getElementText());
             } else if (isStarted == true
                     && eventType == XMLStreamConstants.START_ELEMENT
                     && reader.getName().toString().equals("Description")) {
-                isDescription = true;
-            } else if (isDescription == true
-                    && eventType == XMLStreamConstants.START_ELEMENT
-                    && reader.getName().toString().equals("SampleName")) {
-                bioSampleBean.setName(reader.getElementText());
-            } else if (isDescription == true
-                    && eventType == XMLStreamConstants.START_ELEMENT
-                    && reader.getName().toString().equals("Title")) {
-                bioSampleBean.setTitle(reader.getElementText());
+                sampleBean.setDescription(reader.getElementText());
             } else if (isStarted == true
                     && eventType == XMLStreamConstants.END_ELEMENT
-                    && reader.getName().toString().equals("BioSample")) {
+                    && reader.getName().toString().equals("SAMPLE")) {
                 isStarted = false;
-                bioSampleBeanList.add(bioSampleBean);
+                sampleBeanList.add(sampleBean);
             }
         }
 
         reader.close();
 
-        return bioSampleBeanList;
+        return sampleBeanList;
     }
 }
